Refactoring

+++ src/main/java/com/example/petclinic/controller/PetController.java
@@ -21,7 +21,7 @@ public class PetController implements BasicController<Pet> {
     }
 
     @Override
-    public Pet get(int id) {
+    public Pet get(Long id) {
 
         return this.petService.get(id);
     }
