+++ src/main/java/com/example/petclinic/service/VetService.java
@@ -21,7 +21,7 @@ public class VetService implements BasicService<Vet> {
     }
 
     @Override
-    public Vet get(int id) {
+    public Vet get(Long id) {
 
         return this.vetRepository.read(new Vet(id));
     }
